Thêm giỏ hàng , chỉnh lại món

+++ FAST FOOD/Areas/Admin/Controllers/DanhmucsController.cs
@@ -1,15 +1,18 @@
-﻿using System;
+﻿using FAST_FOOD.Areas.Admin.Filters;
+using FAST_FOOD.Models;
+using System;
 using System.Collections.Generic;
 using System.Data;
 using System.Data.Entity;
+using System.IO;
 using System.Linq;
 using System.Net;
 using System.Web;
 using System.Web.Mvc;
-using FAST_FOOD.Models;
 
 namespace FAST_FOOD.Areas.Admin.Controllers
 {
+    //[AdminAuthorize]
     public class DanhmucsController : Controller
     {
         private KFCContext db = new KFCContext();
@@ -36,56 +39,92 @@ namespace FAST_FOOD.Areas.Admin.Controllers
         }
 
         // GET: Danhmucs/Create
+        // GET: Danhmucs/Create
         public ActionResult Create()
         {
-            //ViewBag.DanhMucId = new SelectList(db.DanhMucs, "DanhMucId", "TenDanhMuc");
-            return View();
+            return View(new FAST_FOOD.Models.Danhmuc());
         }
 
-        // POST: Danhmucs/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to, for 
-        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
+        // POST: Admin/Danhmucs/Create
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include = "DanhMucId,TenDanhMuc")] Danhmuc danhmuc)
+        public ActionResult Create([Bind(Include = "DanhMucId,TenDanhMuc")] Danhmuc danhmuc, HttpPostedFileBase UploadImage)
         {
             if (ModelState.IsValid)
             {
-                //save picture 
-                
+                // Kiểm tra và tạo thư mục nếu chưa có
+                string folderPath = Server.MapPath("~/Images/Products/");
+                if (!Directory.Exists(folderPath))
+                {
+                    Directory.CreateDirectory(folderPath);
+                }
+
+                // Nếu có upload ảnh thì lưu file
+                if (UploadImage != null && UploadImage.ContentLength > 0)
+                {
+                    string fileName = Path.GetFileName(UploadImage.FileName);
+                    string filePath = Path.Combine(Server.MapPath("~/Images/Products/"), fileName);
+
+                    UploadImage.SaveAs(filePath);
+
+                    danhmuc.HinhAnh = "~/Images/Products/" + fileName;
+                    // ✔ CHỈ LƯU TÊN FILE
+                }
+
+                // Lưu xuống DB
                 db.Danhmucs.Add(danhmuc);
                 db.SaveChanges();
+
                 return RedirectToAction("Index");
             }
 
             return View(danhmuc);
         }
-
         // GET: Danhmucs/Edit/5
-        public ActionResult Edit(int? id)
+        [HttpGet]
+        public ActionResult Edit(int id)
         {
-            if (id == null)
-            {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
-            Danhmuc danhmuc = db.Danhmucs.Find(id);
-            if (danhmuc == null)
-            {
-                return HttpNotFound();
-            }
-            return View(danhmuc);
-        }
+            System.Diagnostics.Debug.WriteLine("===== DEBUG INFO =====");
+            System.Diagnostics.Debug.WriteLine("Connection: " + db.Database.Connection.ConnectionString);
+            System.Diagnostics.Debug.WriteLine("Danhmuc count: " + db.Danhmucs.Count());
 
+            var item = db.Danhmucs.Find(id);
+            if (item == null)
+            {
+                System.Diagnostics.Debug.WriteLine("Không tìm thấy ID " + id);
+                TempData["Error"] = $"Không tìm thấy danh mục (ID={id}).";
+                return RedirectToAction("Index");
+            }
+
+            System.Diagnostics.Debug.WriteLine("Tìm thấy danh mục: " + item.TenDanhMuc);
+            return View(item);
+        }
         // POST: Danhmucs/Edit/5
         // To protect from overposting attacks, enable the specific properties you want to bind to, for 
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "DanhMucId,TenDanhMuc")] Danhmuc danhmuc)
+        public ActionResult Edit([Bind(Include = "DanhMucId,TenDanhMuc")] Danhmuc danhmuc, HttpPostedFileBase UploadImage)
         {
             if (ModelState.IsValid)
             {
-                db.Entry(danhmuc).State = EntityState.Modified;
+                var danhmucGoc = db.Danhmucs.Find(danhmuc.DanhMucId);
+                if (danhmucGoc == null)
+                    return HttpNotFound();
+
+                danhmucGoc.TenDanhMuc = danhmuc.TenDanhMuc;
+
+                if (UploadImage != null && UploadImage.ContentLength > 0)
+                {
+                    string fileName = Path.GetFileName(UploadImage.FileName);
+                    string path = Path.Combine(Server.MapPath("~/Images/Products/"), fileName);
+                    UploadImage.SaveAs(path);
+
+                    danhmucGoc.HinhAnh = "/Images/Products/" + fileName;
+
+                }
+
+
                 db.SaveChanges();
                 return RedirectToAction("Index");
             }
@@ -112,19 +151,21 @@ namespace FAST_FOOD.Areas.Admin.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
-            Danhmuc danhmuc = db.Danhmucs.Find(id);
+            var danhmuc = db.Danhmucs.Find(id);
+            if (danhmuc == null)
+                return HttpNotFound();
+
+            bool hasProducts = db.MonAns.Any(p => p.DanhMucId == id);
+            if (hasProducts)
+            {
+                ModelState.AddModelError("", "Không thể xóa danh mục vì đang có sản phẩm thuộc danh mục này!");
+                return View(danhmuc);
+            }
+
             db.Danhmucs.Remove(danhmuc);
             db.SaveChanges();
+
             return RedirectToAction("Index");
         }
-
-        protected override void Dispose(bool disposing)
-        {
-            if (disposing)
-            {
-                db.Dispose();
-            }
-            base.Dispose(disposing);
-        }
     }
-}
+    }
